CLN: gather all decorators into _decorator.py
